examples/sample-dag-project/requirements.txt
apache-airflow==2.9.2
boto3==1.28.57
pandas==2.0.3
pyarrow==12.0.1
sqlalchemy==1.4.48